+++ src/ledger/StorageHelperImpl.cpp
@@ -14,6 +14,8 @@
 #include "SignerHelperImpl.h"
 #include "SignerRuleHelperImpl.h"
 #include "SignerRoleHelperImpl.h"
+#include "VoteHelperImpl.h"
+#include "PollHelperImpl.h"
 #include "LicenseHelperImpl.h"
 #include "StampHelperImpl.h"
 
@@ -40,7 +42,9 @@ StorageHelperImpl::StorageHelperImpl(Database& db, LedgerDelta* ledgerDelta)
         {LedgerEntryType::EXTERNAL_SYSTEM_ACCOUNT_ID_POOL_ENTRY, &getExternalSystemAccountIDPoolEntryHelper()},
         {LedgerEntryType::EXTERNAL_SYSTEM_ACCOUNT_ID, &getExternalSystemAccountIDHelper()},
         {LedgerEntryType::STAMP, &getStampHelper()},
-        {LedgerEntryType::LICENSE, &getLicenseHelper()}
+        {LedgerEntryType::LICENSE, &getLicenseHelper()},
+        {LedgerEntryType::VOTE, &getVoteHelper()},
+        {LedgerEntryType::POLL, &getPollHelper()}
     };
 }
 
@@ -317,6 +321,26 @@ StorageHelperImpl::getStampHelper()
         mStampHelper = std::make_unique<StampHelperImpl>(*this);
     }
     return *mStampHelper;
+}
+
+VoteHelper&
+StorageHelperImpl::getVoteHelper()
+{
+    if(!mVoteHelper)
+    {
+        mVoteHelper = std::make_unique<VoteHelperImpl>(*this);
+    }
+    return *mVoteHelper;
+}
+
+PollHelper&
+StorageHelperImpl::getPollHelper()
+{
+    if(!mPollHelper)
+    {
+        mPollHelper = std::make_unique<PollHelperImpl>(*this);
+    }
+    return *mPollHelper;
 }
 
 } // namespace stellar